More Ore Updates
Difficult choices. It seems that the mods in Lite remove Nikolite and Sapphire from Overworld and restrict it to Nether. So, I decided to convert Sapphire Ore to Ferrous Ore, an ore that was not included in Classic. I also switched Nikolite ore to Silver ore since Factorization adds more silver to the game for producing Lead. The amounts and the layers at which they appear may not be accurate but at least it introduces some new ores into your existing map instead of having to wait for new chunks to generate.